requirements.txt
asgiref==3.7.2
Django==4.2.3
django-filter==23.2
django-js-asset==2.1.0
django-mptt==0.14.0
djangorestframework==3.14.0
drf-yasg==1.21.6
inflection==0.5.1
packaging==23.1
Pillow==10.0.0
PyPDF2==3.0.1
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
uritemplate==4.1.1
